+++ Domain/MetaMat/DenseMat.hpp
@@ -43,7 +43,7 @@ protected:
     const T* const memory = nullptr;
 
 public:
-    using MetaMat<T>::solve;
+    using MetaMat<T>::direct_solve;
 
     DenseMat(uword, uword, uword);
     DenseMat(const DenseMat&);
@@ -55,7 +55,8 @@ public:
     [[nodiscard]] bool is_empty() const override;
     void zeros() override;
 
-    T max() const override;
+    [[nodiscard]] T max() const override;
+    [[nodiscard]] Col<T> diag() const override;
 
     const T* memptr() const override;
     T* memptr() override;
@@ -137,6 +138,14 @@ template<sp_d T> void DenseMat<T>::zeros() {
 
 template<sp_d T> T DenseMat<T>::max() const { return op_max::direct_max(memptr(), this->n_elem); }
 
+template<sp_d T> Col<T> DenseMat<T>::diag() const {
+    Col<T> diag_vec(std::min(this->n_rows, this->n_cols), fill::none);
+
+    suanpan_for(0llu, diag_vec.n_elem, [&](const uword I) { diag_vec(I) = this->operator()(I, I); });
+
+    return diag_vec;
+}
+
 template<sp_d T> const T* DenseMat<T>::memptr() const { return memory; }
 
 template<sp_d T> T* DenseMat<T>::memptr() { return const_cast<T*>(memory); }
@@ -180,6 +189,7 @@ template<sp_d T> void DenseMat<T>::operator-=(const triplet_form<T, uword>& M) {
 template<sp_d T> void DenseMat<T>::operator*=(const T value) { arrayops::inplace_mul(memptr(), value, this->n_elem); }
 
 template<sp_d T> int DenseMat<T>::sign_det() const {
+    if(IterativeSolver::NONE != this->setting.iterative_solver) throw invalid_argument("analysis requires the sign of determinant but iterative solver does not support it");
     auto det_sign = 1;
     for(unsigned I = 0; I < pivot.n_elem; ++I) if((this->operator()(I, I) < 0.) ^ (static_cast<int>(I) + 1 != pivot(I))) det_sign = -det_sign;
     return det_sign;