+++ Domain/MetaMat/FullMat.hpp
@@ -50,10 +50,10 @@ public:
 
     T& at(uword, uword) override;
 
-    Mat<T> operator*(const Mat<T>&) override;
+    Mat<T> operator*(const Mat<T>&) const override;
 
-    int solve(Mat<T>&, Mat<T>&&) override;
-    int solve(Mat<T>&, const Mat<T>&) override;
+    int direct_solve(Mat<T>&, Mat<T>&&) override;
+    int direct_solve(Mat<T>&, const Mat<T>&) override;
 };
 
 template<sp_d T> FullMat<T>::FullMat(const uword in_rows, const uword in_cols)
@@ -80,7 +80,7 @@ template<sp_d T> T& FullMat<T>::at(const uword in_row, const uword in_col) {
     return access::rw(this->memory[in_row + in_col * this->n_rows]);
 }
 
-template<sp_d T> Mat<T> FullMat<T>::operator*(const Mat<T>& B) {
+template<sp_d T> Mat<T> FullMat<T>::operator*(const Mat<T>& B) const {
     Mat<T> C(arma::size(B));
 
     const auto M = static_cast<int>(this->n_rows);
@@ -116,7 +116,7 @@ template<sp_d T> Mat<T> FullMat<T>::operator*(const Mat<T>& B) {
     return C;
 }
 
-template<sp_d T> int FullMat<T>::solve(Mat<T>& X, const Mat<T>& B) {
+template<sp_d T> int FullMat<T>::direct_solve(Mat<T>& X, const Mat<T>& B) {
     if(this->factored) return this->solve_trs(X, B);
 
     auto N = static_cast<int>(this->n_rows);
@@ -131,7 +131,7 @@ template<sp_d T> int FullMat<T>::solve(Mat<T>& X, const Mat<T>& B) {
         X = B;
         arma_fortran(arma_sgesv)(&N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)X.memptr(), &LDB, &INFO);
     }
-    else if(Precision::FULL == this->precision) {
+    else if(Precision::FULL == this->setting.precision) {
         using E = double;
         X = B;
         arma_fortran(arma_dgesv)(&N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)X.memptr(), &LDB, &INFO);
@@ -158,7 +158,7 @@ template<sp_d T> int FullMat<T>::solve_trs(Mat<T>& X, const Mat<T>& B) {
         X = B;
         arma_fortran(arma_sgetrs)(&TRAN, &N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)X.memptr(), &LDB, &INFO);
     }
-    else if(Precision::FULL == this->precision) {
+    else if(Precision::FULL == this->setting.precision) {
         using E = double;
         X = B;
         arma_fortran(arma_dgetrs)(&TRAN, &N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)X.memptr(), &LDB, &INFO);
@@ -171,8 +171,8 @@ template<sp_d T> int FullMat<T>::solve_trs(Mat<T>& X, const Mat<T>& B) {
         auto multiplier = norm(full_residual);
 
         auto counter = 0u;
-        while(counter++ < this->refinement) {
-            if(multiplier < this->tolerance) break;
+        while(counter++ < this->setting.iterative_refinement) {
+            if(multiplier < this->setting.tolerance) break;
 
             auto residual = conv_to<fmat>::from(full_residual / multiplier);
 
@@ -190,7 +190,7 @@ template<sp_d T> int FullMat<T>::solve_trs(Mat<T>& X, const Mat<T>& B) {
     return INFO;
 }
 
-template<sp_d T> int FullMat<T>::solve(Mat<T>& X, Mat<T>&& B) {
+template<sp_d T> int FullMat<T>::direct_solve(Mat<T>& X, Mat<T>&& B) {
     if(this->factored) return this->solve_trs(X, std::forward<Mat<T>>(B));
 
     auto N = static_cast<int>(this->n_rows);
@@ -207,7 +207,7 @@ template<sp_d T> int FullMat<T>::solve(Mat<T>& X, Mat<T>&& B) {
         arma_fortran(arma_sgesv)(&N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)B.memptr(), &LDB, &INFO);
         X = std::move(B);
     }
-    else if(Precision::FULL == this->precision) {
+    else if(Precision::FULL == this->setting.precision) {
         using E = double;
         arma_fortran(arma_dgesv)(&N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)B.memptr(), &LDB, &INFO);
         X = std::move(B);
@@ -234,7 +234,7 @@ template<sp_d T> int FullMat<T>::solve_trs(Mat<T>& X, Mat<T>&& B) {
         arma_fortran(arma_sgetrs)(&TRAN, &N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)B.memptr(), &LDB, &INFO);
         X = std::move(B);
     }
-    else if(Precision::FULL == this->precision) {
+    else if(Precision::FULL == this->setting.precision) {
         using E = double;
         arma_fortran(arma_dgetrs)(&TRAN, &N, &NRHS, (E*)this->memptr(), &N, this->pivot.memptr(), (E*)B.memptr(), &LDB, &INFO);
         X = std::move(B);
@@ -245,8 +245,8 @@ template<sp_d T> int FullMat<T>::solve_trs(Mat<T>& X, Mat<T>&& B) {
         auto multiplier = arma::norm(B);
 
         auto counter = 0u;
-        while(counter++ < this->refinement) {
-            if(multiplier < this->tolerance) break;
+        while(counter++ < this->setting.iterative_refinement) {
+            if(multiplier < this->setting.tolerance) break;
 
             auto residual = conv_to<fmat>::from(B / multiplier);