+++ Domain/MetaMat/csc_form.hpp
@@ -77,12 +77,12 @@ public:
 
     void print() const;
 
-    template<sp_d T2> csc_form<data_t, index_t> operator*(const T2 scalar) {
+    template<sp_d T2> csc_form<data_t, index_t> operator*(const T2 scalar) const {
         csc_form<data_t, index_t> copy = *this;
         return copy *= scalar;
     }
 
-    template<sp_d T2> csc_form<data_t, index_t> operator/(const T2 scalar) {
+    template<sp_d T2> csc_form<data_t, index_t> operator/(const T2 scalar) const {
         csc_form<data_t, index_t> copy = *this;
         return copy /= scalar;
     }
@@ -106,13 +106,13 @@ public:
         return access::rw(bin) = data_t(0);
     }
 
-    Mat<data_t> operator*(const Col<data_t>& in_mat) {
+    Mat<data_t> operator*(const Col<data_t>& in_mat) const {
         Mat<data_t> out_mat = arma::zeros<Mat<data_t>>(in_mat.n_rows, 1);
         for(index_t I = 0; I < n_cols; ++I) for(auto J = col_ptr[I]; J < col_ptr[I + 1]; ++J) out_mat(row_idx[J]) += val_idx[J] * in_mat(I);
         return out_mat;
     }
 
-    Mat<data_t> operator*(const Mat<data_t>& in_mat) {
+    Mat<data_t> operator*(const Mat<data_t>& in_mat) const {
         Mat<data_t> out_mat = arma::zeros<Mat<data_t>>(in_mat.n_rows, in_mat.n_cols);
         for(index_t I = 0; I < n_cols; ++I) for(auto J = col_ptr[I]; J < col_ptr[I + 1]; ++J) out_mat.row(row_idx[J]) += val_idx[J] * in_mat.row(I);
         return out_mat;