+++ Solver/Integrator/GeneralizedAlpha.cpp
@@ -85,8 +85,8 @@ int GeneralizedAlpha::process_load() {
     const auto& D = get_domain().lock();
     auto& W = D->get_factory();
 
-    const auto current_time = W->get_current_time();
-    const auto trial_time = W->get_trial_time();
+    const sp_d auto current_time = W->get_current_time();
+    const sp_d auto trial_time = W->get_trial_time();
 
     W->update_trial_time(F1 * current_time + F2 * trial_time);
 
@@ -101,8 +101,8 @@ int GeneralizedAlpha::process_constraint() {
     const auto& D = get_domain().lock();
     auto& W = D->get_factory();
 
-    const auto current_time = W->get_current_time();
-    const auto trial_time = W->get_trial_time();
+    const sp_d auto current_time = W->get_current_time();
+    const sp_d auto trial_time = W->get_trial_time();
 
     W->update_trial_time(F1 * current_time + F2 * trial_time);
 
@@ -117,8 +117,8 @@ int GeneralizedAlpha::process_load_resistance() {
     const auto& D = get_domain().lock();
     auto& W = D->get_factory();
 
-    const auto current_time = W->get_current_time();
-    const auto trial_time = W->get_trial_time();
+    const sp_d auto current_time = W->get_current_time();
+    const sp_d auto trial_time = W->get_trial_time();
 
     W->update_trial_time(F1 * current_time + F2 * trial_time);
 
@@ -133,8 +133,8 @@ int GeneralizedAlpha::process_constraint_resistance() {
     const auto& D = get_domain().lock();
     auto& W = D->get_factory();
 
-    const auto current_time = W->get_current_time();
-    const auto trial_time = W->get_trial_time();
+    const sp_d auto current_time = W->get_current_time();
+    const sp_d auto trial_time = W->get_trial_time();
 
     W->update_trial_time(F1 * current_time + F2 * trial_time);