+++ Toolbox/LBFGS.hpp
@@ -131,13 +131,6 @@ public:
     }
 };
 
-class Quadratic {
-public:
-    [[nodiscard]] vec evaluate_residual(const vec& x) const { return square(x) - 1.; }
-
-    [[nodiscard]] mat evaluate_jacobian(const vec& x) const { return diagmat(2. * x); }
-};
-
 #endif
 
 //! @}